requirements.txt
openai
tqdm
python-dotenv
pandas
google-genai
flask